+++ src/code.rs
@@ -35,7 +35,7 @@ impl Code {
             r2[0] = c2;
         }
         let mut j = L1CA_CODE_LEN - G2_DELAY[(prn - 1) as usize];
-#[allow(clippy::needless_range_loop)]
+        #[allow(clippy::needless_range_loop)]
         for i in 0..L1CA_CODE_LEN {
             let v = -g1[i] * g2[j % L1CA_CODE_LEN];
             g.push(v);
@@ -59,10 +59,10 @@ impl Code {
         }
     }
 
-    pub fn get_code_len(sig: &str) -> f64 {
+    pub fn get_code_len(sig: &str) -> usize {
         match sig {
-            "L1CA" => L1CA_CODE_LEN as f64,
-            _ => 0.0,
+            "L1CA" => L1CA_CODE_LEN,
+            _ => 0,
         }
     }
